+++ lpmp_py/__init__.py
@@ -1,2 +1,9 @@
-import .graph_matching
-import .multigraph_matching
+from .raw_solvers import gm_solver, mgm_solver
+
+try:
+    import torch
+except ImportError as e:
+    import warnings
+    warnings.warn("Missing Torch, skipping import of torch modules")
+else:
+    from .torch_wrappers import *
