+++ lpmp_py/torch_wrappers/__init__.py
@@ -1,0 +1,2 @@
+from .graph_matching import *
+from .multigraph_matching import *
